+++ src/services/ec2/connections.ts
@@ -16,6 +16,12 @@ import { RawAwsSubnet } from '../subnet/data'
 import { RawAwsEcsContainer } from '../ecsContainer/data'
 import { RawAwsSystemsManagerInstance } from '../systemsManagerInstance/data'
 import { ssmManagedInstanceArn } from '../../utils/generateArns'
+import { RawAwsElasticBeanstalkEnv } from '../elasticBeanstalkEnvironment/data'
+import { RawAwsEksCluster } from '../eksCluster/data'
+import { getEksClusterName, getElasticBeanstalkEnvId } from './utils'
+import { RawAwsInstanceProfile } from '../iamInstanceProfile/data'
+import { globalRegionName } from '../../enums/regions'
+import { RawAwsIamRole } from '../iamRole/data'
 
 /**
  * EC2
@@ -35,6 +41,7 @@ export default ({
     KeyPairName?: string
     Tags?: TagList
     IamInstanceProfile: IamInstanceProfile
+    IamRolesArn?: string[]
   }
   region: string
 }): { [key: string]: ServiceConnection[] } => {
@@ -44,6 +51,9 @@ export default ({
     SecurityGroups: instanceSecurityGroups = [],
     NetworkInterfaces: instanceNetworkInterfaces = [],
     SubnetId: subnetId,
+    Tags: tags,
+    IamInstanceProfile: iamInstanceProfile,
+    IamRolesArn: rolesArn,
   } = instance
 
   /**
@@ -189,17 +199,37 @@ export default ({
           id: subnet.SubnetId,
           resourceType: services.subnet,
           relation: 'child',
-          field: 'subnet',
+          field: 'subnets',
         })
       }
     }
   }
 
   /**
-   * Find EKS
+   * Find EKS cluster
    * related to this EC2
    */
-  // TODO: Implement when eks service is ready
+  const eksClusterName = getEksClusterName(tags)
+  const eksClusters: {
+    name: string
+    data: { [property: string]: any[] }
+  } = data.find(({ name }) => name === services.eksCluster)
+  if (eksClusters?.data?.[region]) {
+    const eksClustersInRegion: RawAwsEksCluster[] = eksClusters.data[
+      region
+    ].filter(({ name }: RawAwsEksCluster) => name === eksClusterName)
+
+    if (!isEmpty(eksClustersInRegion)) {
+      for (const eksCluster of eksClustersInRegion) {
+        connections.push({
+          id: eksCluster.arn,
+          resourceType: services.eksCluster,
+          relation: 'child',
+          field: 'eksCluster',
+        })
+      }
+    }
+  }
 
   /**
    * Find ECS Container
@@ -210,9 +240,9 @@ export default ({
     data: { [property: string]: any[] }
   } = data.find(({ name }) => name === services.ecsContainer)
   if (ecsContainers?.data?.[region]) {
-    const containersInRegion: RawAwsEcsContainer[] = ecsContainers.data[region].filter(
-      ({ ec2InstanceId }) => ec2InstanceId === id
-    )
+    const containersInRegion: RawAwsEcsContainer[] = ecsContainers.data[
+      region
+    ].filter(({ ec2InstanceId }) => ec2InstanceId === id)
 
     if (!isEmpty(containersInRegion)) {
       for (const container of containersInRegion) {
@@ -230,18 +260,24 @@ export default ({
    * Find SSM managed instances
    * related to this EC2 instance
    */
-   const instances: {
+  const instances: {
     name: string
     data: { [property: string]: any[] }
   } = data.find(({ name }) => name === services.systemsManagerInstance)
   if (instances?.data?.[region]) {
-    const dataInRegion: RawAwsSystemsManagerInstance[] = instances.data[region].filter(
+    const dataInRegion: RawAwsSystemsManagerInstance[] = instances.data[
+      region
+    ].filter(
       ({ InstanceId }: RawAwsSystemsManagerInstance) => InstanceId === id
     )
 
     if (!isEmpty(dataInRegion)) {
       for (const ssmInstance of dataInRegion) {
-        const arn = ssmManagedInstanceArn({ region, account, name: ssmInstance.InstanceId })
+        const arn = ssmManagedInstanceArn({
+          region,
+          account,
+          name: ssmInstance.InstanceId,
+        })
         connections.push({
           id: arn,
           resourceType: services.systemsManagerInstance,
@@ -254,9 +290,82 @@ export default ({
 
   /**
    * Find Elastic Beanstalk
-   * related to this EC2 loadbalancer
+   * related to this EC2 instance
    */
-  // TODO: Implement when eb service is ready
+  const elasticBeanstalkEnvId = getElasticBeanstalkEnvId(tags)
+  const elasticBeanstalkEnvs: {
+    name: string
+    data: { [property: string]: any[] }
+  } = data.find(({ name }) => name === services.elasticBeanstalkEnv)
+  if (elasticBeanstalkEnvs?.data?.[region]) {
+    const elasticBeanstalkEnvsInRegion: RawAwsElasticBeanstalkEnv[] =
+      elasticBeanstalkEnvs.data[region].filter(
+        ({ EnvironmentId }: RawAwsElasticBeanstalkEnv) =>
+          elasticBeanstalkEnvId === EnvironmentId
+      )
+
+    if (!isEmpty(elasticBeanstalkEnvsInRegion)) {
+      for (const elasticBeanstalkEnv of elasticBeanstalkEnvsInRegion) {
+        connections.push({
+          id: elasticBeanstalkEnv.EnvironmentId,
+          resourceType: services.elasticBeanstalkEnv,
+          relation: 'child',
+          field: 'elasticBeanstalkEnv',
+        })
+      }
+    }
+  }
+
+  /**
+   * Find IAM Instance Profiles
+   * related to this EC2 instance
+   */
+  const iamInstanceProfiles: {
+    name: string
+    data: { [property: string]: any[] }
+  } = data.find(({ name }) => name === services.iamInstanceProfile)
+  if (iamInstanceProfiles?.data?.[globalRegionName]) {
+    const dataAtRegion: RawAwsInstanceProfile[] = iamInstanceProfiles.data[
+      globalRegionName
+    ].filter(instanceProfile => instanceProfile.Arn === iamInstanceProfile?.Arn)
+    if (!isEmpty(dataAtRegion)) {
+      for (const iamInstance of dataAtRegion) {
+        const { InstanceProfileId: instanceProfileId }: RawAwsInstanceProfile = iamInstance
+
+        connections.push({
+          id: instanceProfileId,
+          resourceType: services.iamInstanceProfile,
+          relation: 'child',
+          field: 'iamInstanceProfile',
+        })
+      }
+    }
+  }
+
+  /**
+   * Find IAM Roles
+   * related to this EC2 instance
+   */
+   const roles: { name: string; data: { [property: string]: any[] } } =
+     data.find(({ name }) => name === services.iamRole)
+ 
+   if (roles?.data?.[globalRegionName]) {
+     const dataAtRegion: RawAwsIamRole[] = roles.data[globalRegionName].filter(
+       ({ Arn }: RawAwsIamRole) => rolesArn?.includes(Arn)
+     )
+     if (!isEmpty(dataAtRegion)) {
+       for (const iamRole of dataAtRegion) {
+         const { Arn: arn } :RawAwsIamRole = iamRole
+ 
+         connections.push({
+           id: arn,
+           resourceType: services.iamRole,
+           relation: 'child',
+           field: 'iamRole',
+         })
+       }
+     }
+   }
 
   const ec2Result = {
     [id]: connections,