+++ src/types/generated.ts
@@ -354,6 +354,7 @@ export type AwsAppSync = AwsBaseService & {
   dataSources?: Maybe<Array<Maybe<AwsAppSyncDataSource>>>;
   dynamodb?: Maybe<Array<Maybe<AwsDynamoDbTable>>>;
   functions?: Maybe<Array<Maybe<AwsAppSyncFunction>>>;
+  iamRoles?: Maybe<Array<Maybe<AwsIamRole>>>;
   lambda?: Maybe<Array<Maybe<AwsLambda>>>;
   lambdaAuthorizerIdentityValidationExpression?: Maybe<Scalars['String']>;
   lambdaAuthorizerResultTtlInSeconds?: Maybe<Scalars['Int']>;
@@ -375,6 +376,7 @@ export type AwsAppSync = AwsBaseService & {
   userPoolDefaultAction?: Maybe<Scalars['String']>;
   userPoolId?: Maybe<Scalars['String']>;
   wafWebAclArn?: Maybe<Scalars['String']>;
+  webAcl?: Maybe<Array<Maybe<AwsWafV2WebAcl>>>;
   xrayEnabled?: Maybe<Scalars['String']>;
 };
 
@@ -1460,11 +1462,14 @@ export type AwsEc2 = AwsBaseService & {
   ebsOptimized?: Maybe<Scalars['String']>;
   ecsContainer?: Maybe<Array<Maybe<AwsEcsContainer>>>;
   eip?: Maybe<Array<Maybe<AwsEip>>>;
+  eksCluster?: Maybe<Array<Maybe<AwsEksCluster>>>;
+  elasticBeanstalkEnv?: Maybe<Array<Maybe<AwsElasticBeanstalkEnv>>>;
   elasticIps?: Maybe<Scalars['String']>;
   emrInstance?: Maybe<Array<Maybe<AwsEmrInstance>>>;
   ephemeralBlockDevices?: Maybe<Array<Maybe<AwsEc2Blockdevice>>>;
   hibernation?: Maybe<Scalars['String']>;
-  iamInstanceProfile?: Maybe<Scalars['String']>;
+  iamInstanceProfile?: Maybe<Array<Maybe<AwsIamInstanceProfile>>>;
+  iamRole?: Maybe<Array<Maybe<AwsIamRole>>>;
   instanceLifecycle?: Maybe<Scalars['String']>;
   instanceState?: Maybe<Scalars['String']>;
   instanceType?: Maybe<Scalars['String']>;
@@ -1486,7 +1491,7 @@ export type AwsEc2 = AwsBaseService & {
   securityGroupIds?: Maybe<Array<Maybe<Scalars['String']>>>;
   securityGroups?: Maybe<Array<Maybe<AwsSecurityGroup>>>;
   sourceDestCheck?: Maybe<Scalars['String']>;
-  subnet?: Maybe<Array<Maybe<AwsSubnet>>>;
+  subnets?: Maybe<Array<Maybe<AwsSubnet>>>;
   systemsManagerInstance?: Maybe<Array<Maybe<AwsSystemsManagerInstance>>>;
   tags?: Maybe<Array<Maybe<AwsRawTag>>>;
   tenancy?: Maybe<Scalars['String']>;
@@ -1622,8 +1627,8 @@ export type AwsEcsContainer = AwsBaseService & {
   attachments?: Maybe<Array<Maybe<AwsEcsAttachment>>>;
   attributes?: Maybe<Array<Maybe<AwsEcsAttribute>>>;
   capacityProviderName?: Maybe<Scalars['String']>;
-  ec2Instance?: Maybe<Array<Maybe<AwsEc2>>>;
   ec2InstanceId?: Maybe<Scalars['String']>;
+  ec2Instances?: Maybe<Array<Maybe<AwsEc2>>>;
   ecsTask?: Maybe<Array<Maybe<AwsEcsTask>>>;
   pendingTasksCount?: Maybe<Scalars['Int']>;
   registeredAt?: Maybe<Scalars['String']>;
@@ -2230,6 +2235,7 @@ export type AwsEksCluster = AwsBaseService & {
   certificateAuthority?: Maybe<AwsEksCertificate>;
   clientRequestToken?: Maybe<Scalars['String']>;
   createdAt?: Maybe<Scalars['String']>;
+  ec2Instances?: Maybe<Array<Maybe<AwsEc2>>>;
   encryptionConfig?: Maybe<Array<Maybe<AwsEksEncryptionConfig>>>;
   endpoint?: Maybe<Scalars['String']>;
   iamRoles?: Maybe<Array<Maybe<AwsIamRole>>>;
@@ -2242,7 +2248,7 @@ export type AwsEksCluster = AwsBaseService & {
   resourcesVpcConfig?: Maybe<AwsEksVpcConfigResponse>;
   securityGroups?: Maybe<Array<Maybe<AwsSecurityGroup>>>;
   status?: Maybe<Scalars['String']>;
-  subnet?: Maybe<Array<Maybe<AwsSubnet>>>;
+  subnets?: Maybe<Array<Maybe<AwsSubnet>>>;
   tags?: Maybe<Array<Maybe<AwsRawTag>>>;
   version?: Maybe<Scalars['String']>;
   vpc?: Maybe<Array<Maybe<AwsVpc>>>;
@@ -2487,6 +2493,7 @@ export type AwsElasticBeanstalkEnv = AwsBaseService & {
   applicationName?: Maybe<Scalars['String']>;
   cname?: Maybe<Scalars['String']>;
   description?: Maybe<Scalars['String']>;
+  ec2Instances?: Maybe<Array<Maybe<AwsEc2>>>;
   elasticBeanstalkApp?: Maybe<Array<Maybe<AwsElasticBeanstalkApp>>>;
   endpointUrl?: Maybe<Scalars['String']>;
   name?: Maybe<Scalars['String']>;
@@ -2960,6 +2967,7 @@ export type AwsIamGroup = AwsBaseService & {
 
 export type AwsIamInstanceProfile = AwsBaseService & {
   createDate?: Maybe<Scalars['DateTime']>;
+  ec2Instances?: Maybe<Array<Maybe<AwsEc2>>>;
   iamRole?: Maybe<Array<Maybe<AwsIamRole>>>;
   name?: Maybe<Scalars['String']>;
   path?: Maybe<Scalars['String']>;
@@ -3034,6 +3042,7 @@ export type AwsIamPolicy = AwsBaseService & {
 };
 
 export type AwsIamRole = AwsBaseService & {
+  appSync?: Maybe<Array<Maybe<AwsAppSync>>>;
   assumeRolePolicy?: Maybe<AwsIamJsonPolicy>;
   cloudFormationStack?: Maybe<Array<Maybe<AwsCloudFormationStack>>>;
   codebuilds?: Maybe<Array<Maybe<AwsCodebuild>>>;
@@ -3042,6 +3051,7 @@ export type AwsIamRole = AwsBaseService & {
   createdAt?: Maybe<Scalars['String']>;
   description?: Maybe<Scalars['String']>;
   dynamodb?: Maybe<Array<Maybe<AwsDynamoDbTable>>>;
+  ec2Instances?: Maybe<Array<Maybe<AwsEc2>>>;
   ecsService?: Maybe<Array<Maybe<AwsEcsService>>>;
   eksCluster?: Maybe<Array<Maybe<AwsEksCluster>>>;
   flowLogs?: Maybe<Array<Maybe<AwsFlowLog>>>;
@@ -3887,7 +3897,7 @@ export type AwsSubnet = AwsBaseService & {
   codebuilds?: Maybe<Array<Maybe<AwsCodebuild>>>;
   defaultForAz?: Maybe<Scalars['Boolean']>;
   dmsReplicationInstances?: Maybe<Array<Maybe<AwsDmsReplicationInstance>>>;
-  ec2Instance?: Maybe<Array<Maybe<AwsEc2>>>;
+  ec2Instances?: Maybe<Array<Maybe<AwsEc2>>>;
   ecsService?: Maybe<Array<Maybe<AwsEcsService>>>;
   efsMountTarget?: Maybe<Array<Maybe<AwsEfsMountTarget>>>;
   eksCluster?: Maybe<Array<Maybe<AwsEksCluster>>>;
@@ -4436,6 +4446,7 @@ export type AwsWafV2VisibilityConfig = {
 
 export type AwsWafV2WebAcl = AwsBaseService & {
   ManagedByFirewallManager?: Maybe<Scalars['Boolean']>;
+  appSync?: Maybe<Array<Maybe<AwsAppSync>>>;
   capacity?: Maybe<Scalars['Int']>;
   cloudfront?: Maybe<Array<Maybe<AwsCloudfront>>>;
   customResponseBodies?: Maybe<Array<Maybe<AwsWafV2CustomResponseBody>>>;